+++ app/_actions/category.ts
@@ -5,6 +5,8 @@ import connectToDB from "../_configs/database"
 import { SHOW_IN_PAGE } from "../_constants/gobalVariables"
 import categoryModel from "../_models/category.module"
 import { messageCreator } from "../_utils/messageCreator"
+import { ICategory } from "../_types/category"
+import isAdmin from "../_middlewares/isAdmin"
 
 interface IGetCategoriesProps {
     page: number,
@@ -18,8 +20,9 @@ interface IInsertCategory {
 }
 
 export const getCategories = async ({ page, sort = 'asc' }: IGetCategoriesProps) => {
+    await connectToDB()
 
-    const categories = await categoryModel
+    const categories: ICategory[] = await categoryModel
         .find({})
         .sort({ _id: sort === 'asc' ? -1 : 1 }).skip((page - 1) * SHOW_IN_PAGE).limit(page * SHOW_IN_PAGE)
 
@@ -31,12 +34,7 @@ export const getCategories = async ({ page, sort = 'asc' }: IGetCategoriesProps)
     const categoriesActiveCount = categories.length;
     const categoriesNonActiveCount = 0
 
-    const categoriesDetails: {
-        categoriesCount: number,
-        categoriesAllCount: number,
-        categoriesActiveCount: number,
-        categoriesNonActiveCount: number
-    } = {
+    const categoriesDetails = {
         categoriesCount,
         categoriesAllCount,
         categoriesActiveCount,
@@ -48,12 +46,19 @@ export const getCategories = async ({ page, sort = 'asc' }: IGetCategoriesProps)
 
 export const deleteCategory = async ({ categoryID }: { categoryID: string }) => {
     await connectToDB();
+    const isAdminUser = await isAdmin();
+    if (!isAdminUser) return messageCreator(false, 'در حالت تستی امکان حذف نیست')
+
     await categoryModel.findOneAndDelete({ _id: categoryID }).lean();
     revalidatePath('/categories')
     return messageCreator(true, 'دسته بندی حذف شد')
 }
 
 export const insertCategory = async ({ title, titleEn, href }: IInsertCategory) => {
+    await connectToDB()
+    const isAdminUser = await isAdmin();
+    if (!isAdminUser) return messageCreator(false, 'در حالت تستی امکان اضافه کردن نیست')
+
     const isHasBefore = await categoryModel.findOne({ href }).lean();
     if (isHasBefore) return messageCreator(false, 'لینک دسته بندی تکراریه')
     await categoryModel.create({ title, titleEn, href, caption: '' });
@@ -62,7 +67,6 @@ export const insertCategory = async ({ title, titleEn, href }: IInsertCategory) 
 }
 
 export const getAllCategories = async () => {
-    console.log('GGGGGEEEEEEEEET')
     await connectToDB();
     const categories = await categoryModel.find({}).select('title');
     return JSON.parse(JSON.stringify(categories))
